added a lot of work on one shot -- please help!
one shot is not working, it seems that
I will need to set the pulseIn before the
timer goes off.

On the other hand I know that the one shot
is going off, as I can easily place an
led there.

I could probably place another arduino there
just to see if it would work though